requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
python-magic==0.4.24
